Moved Test page parts to the separate files. Moved primary to root slot. Change name from trailingIcon to endIcon
