+++ packages/components/MenuButton/src/MenuButton.types.ts
@@ -7,7 +7,6 @@ export const MenuButtonName = 'MenuButton';
 
 export interface MenuButtonContext {
   showContextualMenu?: boolean;
-  primary?: boolean;
 }
 
 export interface MenuButtonState {
